pyproject.toml
[tool.poetry]
name = "blobbber-backend"
version = "0.1.0"
description = ""
authors = ["Samuel Baetz <samuelbaetz105@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
Flask-RESTful = "^0.3.8"
Flask-SQLAlchemy = "^2.5.1"
psycopg2-binary = "^2.8.6"
Flask-Migrate = "^2.7.0"
Flask-JWT-Extended = "^4.1.0"
azure-storage-blob = "^12.8.0"
azure-identity = "^1.5.0"
Flask-Cors = "^3.0.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
